Erease patch of org-roam-node-visit.
Looking back, I understand how I was thinking when I made this
decision but thinking now I think this was a dumb decision.
First I have formed the habit of using prefixes, second I stopped
opening files because the opened a new window and third I there are
times where I wanted to open a heading in a new frame. Using prefixes
is the only way I see of living.